add disconnect chat user to logout & add params to creating chat user
